Moved getUserCreatedDatasetList and getUserCreatedDatasetCount function to the DatasetStatisticsUtils file
